Move `HostListItem` to `Icinga\Module\Icingadb\Widget\ItemList`
